bugfix: prevent accidental removal of temporary file or directory by subprocess

+++ pylib/temp.py
@@ -1,3 +1,8 @@
+"""
+Safely create temporary file or directory which is automatically removed
+when the object is dereferenced (by the same process that created it,
+not a subprocess)
+"""
 import os
 import tempfile
 import shutil
@@ -7,22 +12,21 @@ class TempFile(file):
         fd, path = tempfile.mkstemp(suffix, prefix)
         os.close(fd)
         self.path = path
+        self.pid = os.getpid()
         file.__init__(self, path, "w")
 
     def __del__(self):
-        os.remove(self.path)
+        if self.pid == os.getpid():
+            os.remove(self.path)
 
 class TempDir:
     def __init__(self, prefix='tmp', suffix='', dir=None):
         self.path = tempfile.mkdtemp(suffix, prefix, dir)
+        self.pid = os.getpid()
 
     def remove(self):
         shutil.rmtree(self.path)
         
     def __del__(self):
-        self.remove()
-
-        
-        
-        
-
+        if self.pid == os.getpid():
+            self.remove()
